Consolidate duplicate code and fix efficiency issues
Extract shared JSON markdown-stripping helper, remove duplicate extension
variable, batch UpdateHasUnmatchedProducts calls, and fix pre-existing
discount test assertion.
